+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -418,7 +418,7 @@ class CreateDatasourceResponseBody(BaseModel):
     metadata: CreateDatasourceMetadata
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KYD71GMF31E1N4SW7DRS99BY"
+        "01KYDMWD70517XDDMCJ9X5RPYP"
     )
     r"""The unique identifier of the data source"""