+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -247,7 +247,7 @@ class ListDatasourcesData(BaseModel):
     metadata: ListDatasourcesMetadata
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KYD71GMAZ4S6A7E74A2JJ5P4"
+        "01KYDMWD6TC07CBJHHRP29S6JP"
     )
     r"""The unique identifier of the data source"""