+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -169,7 +169,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     metadata: RetrieveDatasourceMetadata
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KYD71GMDGV66Q3JX6MFPAW3M"
+        "01KYDMWD6YMK4BGBF948G8NNZH"
     )
     r"""The unique identifier of the data source"""