+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -183,7 +183,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     metadata: UpdateDatasourceMetadata
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KYD71GMKS34235B8W37QDBX8"
+        "01KYDMWD74RWYDYFKJA5J3033F"
     )
     r"""The unique identifier of the data source"""