+++ packages/orq-rc/RELEASES.md
@@ -9729,3 +9729,13 @@ Based on:
 - [python v4.13.0-rc.4] packages/orq-rc
 ### Releases
 - [PyPI v4.13.0-rc.4] https://pypi.org/project/orq-ai-sdk/4.13.0-rc.4 - packages/orq-rc
+
+## 2026-07-25 22:05:48
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.791.0 (2.924.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.13.0-rc.5] packages/orq-rc
+### Releases
+- [PyPI v4.13.0-rc.5] https://pypi.org/project/orq-ai-sdk/4.13.0-rc.5 - packages/orq-rc